Resolve the issues of codestyle of nacos-config module for phase2 (#3253)
* [ISSUE#3249]Resolve the part issues of codestyle for nacos-config module(package:service/dump)


* [ISSUE#3249]Resolve the part issues of codestyle for nacos-config module(package:service/merge)
