+++ config/src/main/java/com/alibaba/nacos/config/server/service/dump/processor/DumpProcessor.java
@@ -1,0 +1,93 @@
+/*
+ * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.nacos.config.server.service.dump.processor;
+
+import com.alibaba.nacos.config.server.manager.AbstractTask;
+import com.alibaba.nacos.config.server.manager.TaskProcessor;
+import com.alibaba.nacos.config.server.model.ConfigInfo;
+import com.alibaba.nacos.config.server.model.ConfigInfo4Beta;
+import com.alibaba.nacos.config.server.model.ConfigInfo4Tag;
+import com.alibaba.nacos.config.server.model.event.ConfigDumpEvent;
+import com.alibaba.nacos.config.server.service.dump.DumpConfigHandler;
+import com.alibaba.nacos.config.server.service.dump.DumpService;
+import com.alibaba.nacos.config.server.service.dump.task.DumpTask;
+import com.alibaba.nacos.config.server.service.repository.PersistService;
+import com.alibaba.nacos.config.server.utils.GroupKey2;
+import org.apache.commons.lang3.StringUtils;
+
+import java.util.Objects;
+
+/**
+ * dump processor.
+ *
+ * @author Nacos
+ * @date 2020/7/5 12:19 PM
+ */
+public class DumpProcessor implements TaskProcessor {
+    
+    public DumpProcessor(DumpService dumpService) {
+        this.dumpService = dumpService;
+    }
+    
+    @Override
+    public boolean process(String taskType, AbstractTask task) {
+        final PersistService persistService = dumpService.getPersistService();
+        DumpTask dumpTask = (DumpTask) task;
+        String[] pair = GroupKey2.parseKey(dumpTask.getGroupKey());
+        String dataId = pair[0];
+        String group = pair[1];
+        String tenant = pair[2];
+        long lastModified = dumpTask.getLastModified();
+        String handleIp = dumpTask.getHandleIp();
+        boolean isBeta = dumpTask.isBeta();
+        String tag = dumpTask.getTag();
+        
+        ConfigDumpEvent.ConfigDumpEventBuilder build = ConfigDumpEvent.builder().namespaceId(tenant).dataId(dataId)
+                .group(group).isBeta(isBeta).tag(tag).lastModifiedTs(lastModified).handleIp(handleIp);
+        
+        if (isBeta) {
+            // beta发布，则dump数据，更新beta缓存
+            ConfigInfo4Beta cf = persistService.findConfigInfo4Beta(dataId, group, tenant);
+            
+            build.remove(Objects.isNull(cf));
+            build.betaIps(Objects.isNull(cf) ? null : cf.getBetaIps());
+            build.content(Objects.isNull(cf) ? null : cf.getContent());
+            
+            return DumpConfigHandler.configDump(build.build());
+        } else {
+            if (StringUtils.isBlank(tag)) {
+                ConfigInfo cf = persistService.findConfigInfo(dataId, group, tenant);
+                
+                build.remove(Objects.isNull(cf));
+                build.content(Objects.isNull(cf) ? null : cf.getContent());
+                build.type(Objects.isNull(cf) ? null : cf.getType());
+                
+                return DumpConfigHandler.configDump(build.build());
+            } else {
+                
+                ConfigInfo4Tag cf = persistService.findConfigInfo4Tag(dataId, group, tenant, tag);
+                
+                build.remove(Objects.isNull(cf));
+                build.content(Objects.isNull(cf) ? null : cf.getContent());
+                
+                return DumpConfigHandler.configDump(build.build());
+            }
+        }
+    }
+    
+    final DumpService dumpService;
+}
