+++ config/src/main/java/com/alibaba/nacos/config/server/service/merge/MergeTaskProcessor.java
@@ -38,13 +38,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Merge task processor
+ * Merge task processor.
  *
  * @author Nacos
  */
 public class MergeTaskProcessor implements TaskProcessor {
     
-    final int PAGE_SIZE = 10000;
+    private static final int PAGE_SIZE = 10000;
     
     MergeTaskProcessor(PersistService persistService, MergeDatumService mergeService) {
         this.persistService = persistService;
@@ -68,33 +68,32 @@ public class MergeTaskProcessor implements TaskProcessor {
                         .findConfigInfoAggrByPage(dataId, group, tenant, pageNo, PAGE_SIZE);
                 if (page != null) {
                     datumList.addAll(page.getPageItems());
-                    log.info("[merge-query] {}, {}, size/total={}/{}", dataId, group, datumList.size(), rowCount);
+                    LOGGER.info("[merge-query] {}, {}, size/total={}/{}", dataId, group, datumList.size(), rowCount);
                 }
             }
             
             final Timestamp time = TimeUtils.getCurrentTime();
-            // 聚合
             if (datumList.size() > 0) {
+                // merge
                 ConfigInfo cf = merge(dataId, group, tenant, datumList);
                 
                 persistService.insertOrUpdate(null, null, cf, time, null);
                 
-                log.info("[merge-ok] {}, {}, size={}, length={}, md5={}, content={}", dataId, group, datumList.size(),
+                LOGGER.info("[merge-ok] {}, {}, size={}, length={}, md5={}, content={}", dataId, group, datumList.size(),
                         cf.getContent().length(), cf.getMd5(), ContentUtils.truncateContent(cf.getContent()));
                 
                 ConfigTraceService
                         .logPersistenceEvent(dataId, group, tenant, null, time.getTime(), InetUtils.getSelfIp(),
                                 ConfigTraceService.PERSISTENCE_EVENT_MERGE, cf.getContent());
-            }
-            // 删除
-            else {
+            } else {
+                // remove
                 if (StringUtils.isBlank(tag)) {
                     persistService.removeConfigInfo(dataId, group, tenant, clientIp, null);
                 } else {
                     persistService.removeConfigInfoTag(dataId, group, tenant, tag, clientIp, null);
                 }
                 
-                log.warn("[merge-delete] delete config info because no datum. dataId=" + dataId + ", groupId=" + group);
+                LOGGER.warn("[merge-delete] delete config info because no datum. dataId=" + dataId + ", groupId=" + group);
                 
                 ConfigTraceService
                         .logPersistenceEvent(dataId, group, tenant, null, time.getTime(), InetUtils.getSelfIp(),
@@ -105,12 +104,21 @@ public class MergeTaskProcessor implements TaskProcessor {
             
         } catch (Exception e) {
             mergeService.addMergeTask(dataId, group, tenant, mergeTask.getClientIp());
-            log.info("[merge-error] " + dataId + ", " + group + ", " + e.toString(), e);
+            LOGGER.info("[merge-error] " + dataId + ", " + group + ", " + e.toString(), e);
         }
         
         return true;
     }
     
+    /**
+     * merge datumList {@link ConfigInfoAggr}.
+     *
+     * @param dataId data id
+     * @param group group
+     * @param tenant tenant
+     * @param datumList datumList
+     * @return {@link ConfigInfo}
+     */
     public static ConfigInfo merge(String dataId, String group, String tenant, List<ConfigInfoAggr> datumList) {
         StringBuilder sb = new StringBuilder();
         String appName = null;
@@ -125,9 +133,7 @@ public class MergeTaskProcessor implements TaskProcessor {
         return new ConfigInfo(dataId, group, tenant, appName, content);
     }
     
-    // =====================
-    
-    private static final Logger log = LoggerFactory.getLogger(MergeTaskProcessor.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(MergeTaskProcessor.class);
     
     private PersistService persistService;
